ai_models/cfd_optimizer/rl_loop.py
import gym
import numpy as np
from stable_baselines3 import PPO
from stable_baselines3.common.env_checker import check_env
import subprocess
import os

class OpenFOAMEnv(gym.Env):
    def __init__(self, case_dir='../../simulations/openfoam', max_steps=100):
        super().__init__()
        self.case_dir = case_dir
        self.max_steps = max_steps
        self.action_space = gym.spaces.Box(low=-1, high=1, shape=(4,), dtype=np.float32)  # Example: 4 morph params
        self.observation_space = gym.spaces.Box(low=-np.inf, high=np.inf, shape=(10,), dtype=np.float32)  # Example: 10 features
        self.current_step = 0
    def reset(self):
        self.current_step = 0
        obs = np.zeros(self.observation_space.shape)
        return obs
    def step(self, action):
        self.current_step += 1
        # Write action to OpenFOAM input files
        self._write_action_to_case(action)
        # Run OpenFOAM simulation
        self._run_openfoam()
        # Parse results
        obs, reward = self._parse_results()
        done = self.current_step >= self.max_steps
        info = {}
        return obs, reward, done, info
    def _write_action_to_case(self, action):
        # Placeholder: Write morph params to OpenFOAM case
        pass
    def _run_openfoam(self):
        # Placeholder: Run OpenFOAM via shell script
        subprocess.run(['bash', 'ai_models/cfd_optimizer/openfoam_runner.sh', self.case_dir], check=True)
    def _parse_results(self):
        # Placeholder: Parse drag, stability from OpenFOAM output
        drag = np.random.uniform(0, 1)
        stability = np.random.uniform(0, 1)
        obs = np.random.randn(10)
        reward = -drag + 0.5 * stability
        return obs, reward

if __name__ == '__main__':
    env = OpenFOAMEnv()
    check_env(env)
    model = PPO('MlpPolicy', env, verbose=1)
    model.learn(total_timesteps=10000)
    model.save('ppo_openfoam') 