requirements.txt
# Core Python & Utilities
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
joblib>=1.2.0

# Deep Learning & ML
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
diffusers>=0.18.0
scikit-learn>=1.3.0
pytorch-lightning>=2.0.0
tensorboard>=2.13.0

# TFLite model export and inference
tensorflow>=2.13.0
tensorflow-cpu>=2.13.0; platform_system!="Linux" or platform_machine!="aarch64"
tensorflow-lite>=2.13.0
tflite-runtime>=2.13.0; platform_system=="Linux" and platform_machine=="aarch64"

# Reinforcement Learning & Simulation
stable-baselines3>=2.1.0
gymnasium>=0.29.0
openai-gym>=0.26.0
pyyaml>=6.0

# Flask Backend & API
flask>=2.3.0
flask-cors>=4.0.0
flask-restful>=0.3.10
gunicorn>=21.2.0

# MQTT & Communication
paho-mqtt>=1.6.1
python-socketio>=5.8.0
requests>=2.31.0

# Serial Communication (Arduino Bridge)
pyserial>=3.5

# Edge AI & Jetson/RPi
opencv-python>=4.8.0
imutils>=0.5.4

# Telemetry Logging & Monitoring
influxdb-client>=1.38.0
psutil>=5.9.5

# Unity & Digital Twin (MQTT Bridge)
unitypack>=0.0.4
protobuf>=4.23.0

# Miscellaneous
notebook>=7.0.0
jupyterlab>=4.0.0
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0

# Optional: For OpenFOAM Python Integration
pyfoam>=0.6.9 